valid parintheses
This solution was developed by one of the other LeetCode users. It is a clever one where they used a built in java replace function to see if any of the bracket pairs are found (), {}, [] and replaces these with “” strings. The loop runs until all the valid pairs are replaced.

Lastly, it checks to see if the string is now empty or not.